Inputs, Ogg support, Check for Focus for mouse. Multi FF8DIR list. More
I added some Input Code. Mostly the ability to bind more than one input to a function.
Ogg support is i'd say 70% working as is. Might be a quality loss from converting from 32bit float to 16bit pcm. LOOPSTART is not working correctly. Should be the point where the SoundEffectInstance starts the loop from. But for some reason it is starting over. It might be working sometimes. I noticed the function gliched when I set a LOOPLENGTH and threw an exception for no reason. Setting LOOPLENGTH to 0 got past that issue. I look forward to FFMPEG replacing the need for SoundEffect.
Focus for mouse. I didn't like how when I was in the MAPs i couldn't tab out and use my mouse. So now when there is no focus the mouse is free. Though this isn't perfect. The mouse could escape.
Multi FF8DIR added a const array of directorys for FF8. When you first look for FF8DIR it will go through the list and pick the first one that exists. I got tired of changing the varible going back and forth from linux.
I took out the GetUnixFullPath(string pt) from some of the functions. I was having issues. Coulda just unset _WINDOWS. I tried to put it back. But everything seems to be working correctly.
Added a IsLinux property to MakiExtended as another way to check for linux with out having to remember to toggle _WINDOWS.
for some reason the field screen isn't working as far as I can tell it isn't anything I changed. I pulled the source down to another folder and it did the same thing.
I wanted to commit before I started trying anything else.