tweaks to menu code added movie player.
I restructured some of the menu code mostly with goal of adding things to it easy.
Like the drawing goes through an array instead of having each thing listed.
added the movie player to the debug menu.
added the intro movie to the new game to play before going to the field.