sound works with out writing files to hard drive.
added a bool to allow enabling of of the encoder else it just loads into ram and skips it. used visual studio to clean up some formatting

though i still make the pcms and load a header into them lol

making the encoder optional isn't pretty  alot of if's added.