created a dictionary to store soundeffects as they are loaded. to prevent leaking memory by creating new soundeffects over and over. and reducing the ammount of loading.
I had it just disposing the sound before it creates new. but then you cannot play more than one sound at a time.

Current version of movie and music sound cannot play at the same time. music must be a dynamicsoundeffectinterface. but video can be a soundeffectinstance.

could be the disposing of the memorystream. but the audio was still playing so i donno why it would be.

I might roll stuff back donno i put music and movie sound in two different interfaces but one still overrides the other. *shrugs*