getting ready to try to get linux to work again
I was testing 64 bit out because I think in linux to work i'll need to run the code in 64 bit. 64 bit doesn't work in windows because the midi library doesn't work. but in linux we can't use it anyway.

i moved 32 bit ffmpeg to x86 folder and 64 bit to x64 folder. i kinda wish it only copied the one it needed unsure how to do that.