CTD linux fixed
//ffmpeg.av_free(Codec); //CTD on linux

maybe it's not a big deal i was trying to free all pointers for memory leaks but linux wasn't happy with this one.

no video in linux but i wonder if that's a formatting thing or something.

ffmpeg.autogen guy is gonna integrate my suggestion in his next version.

removed the old windows only checks for ffmpeg as they were only there to skip over it when the libraries were missing.