Tweaked some settings
It's running on linux video is blank right now got some ideas.

audio was skipping i inceased the buffer size. though this inceased the delay on audio playback (music and video sound). I think the skipping was mostly because i am running on a vm and not a native install of linux.