linux is mostly working minus video image
There is an issue with video in linux. But it'll probably wait till monday. It's hard to see using mono devolop what i'm getting varible wise. On monday I want to try to skip the sws scaler to see if that is what is broken. or if it's something upstream.