new _X64 for 64bit
I added _X64  for any code that requires 64 bit. So we can ise !_X64 for anything that doesn't work on 64 bit.

For some reason DEBUG compiler flag won't work with Debug. So I copied the DebugLinux and set the _WINDOWS flag on it. I was just using that and setting _WINDOWS when I wasn't compiling on Linux.

Atleast for me the DEBUG checkbox would keep getting auto unchecked and the code would stay grayed out.