It's rudimentry mouse controls for menu. :D
so on menus i have it unhide cursor and i added a return rectangle to the draw text function. so i would know where the text was drawn. so then in update check for most position and check each item to see if it's there.
even has mouse wheel control the changing of songs and such when moused over them.