Linux movie audio playback improved. made debug menu slide in
Increased GoalBufferCount to 100;
Added a NextAsync() it runs Next if not Ahead in it's own thread. so less likely to be delayed by the video frames.
I found that via math 100 is about 1 second of audio.

possible to use same functions to put video frames in their own thread. this has not been tested yet.
music could go in a thread too.

I started to do a commit then i got destracted lol

I fixed some infinate loop issues in drawing icons.

I just learned that vectors existed they could simplify code in some places. And rectangles can Inflate. I was doing it all manually. :P

Debug menu slides in from side. I wanted to be able to do this because the loading menu let you slide parts on or off the right or left of screen.

I tweaked the debug menu to be positioned by the top left corner of the text + offset.