I had issue with TEX not working with hires version of textures. So I added code to fall back to classic texture.
I added GetBuffer to TEX to export problem files to temp folder so i try to see what is different about these files.

The high res textures are working with png files in mods.

Have not tested icons.cs yet. It might work. The icons.tex has pcs in multiple files of a image so it could be up to 4x draw calls on those.

Moved TextureHandler into it's own file.