I had to tweak texture handler and main menu.
Had scaling issues i probably created. it was growing to big when  increasing size. and getting too small when shrinking. Gotta be careful to only apply scaling vector to something once.  it can get out of hand lol.

Tweaks to texture handler to correctly handle texture groups that don't have an original complete texture. So it will hold Start00.tex and Start01.tex and draw them correctly. Right now it only works with MENU stuff but should be an "easy" thing to adapt it to use other archives.