progess made on load screen
well it's been a while since I commited lol

I got a basic layout of the select slot screen drawing. I had muliple bugs with entrygroup drawing that I didn't see till I did this. they are now fixed.

the background of windows now only stretches if it's smaller then the desired draw location. in order for all the windows to look the same i had to also set scale instead of auto detecting it.

I added a new value to entry for enabling fill on a texture.

moved more menu files into the menu folder.

I switched over to the newer hires fonts in the menus the low res font doesn't scale well.  I also added a small drop shadow. maybe we can make it optional in the future for people to just swap which font they want on the fly.

added a modded bool to texturehandler so we can detect if modded textures are being used. so i could skip the drawshadow if there is a mod.