starting game selection screen
got a load bar screen working. though it's for looks as it just animates filling the bar and moves on. I also have a loadgame selection screen started. debating how do do it. I need 30 blocks. 3 per screen 10 screens. need to be able to slide it in from side to side.