fixed kernel strings i think. added name function to string that knows where to get the names using faces.ID to id which name is which.
added more to load game class to upto and including Griever's name. using wiki as the guide.

put a state varible in memory to hold the current savegame data. idea being you can alter/read the data needed while playing and can save into the file with some future serialzer? saving will be a bit more techinical as it requres hashing and such to be compatible with steam.

now to make the names read from string be used to instead of hardcoded names.