fixed code on linux
Worked around missing savegame folder on linux. It detects your documents folder but if it doesn't see the folder with the saves in it. it just skips it. so the load screen will be blank.

on thing to note if you have more than one steam id on a system there will be more than 1 folder of saves. so might be an issue.