I kinda got it scaling like I wanted... though it's alittle odd.
I wanted the menu to be able to adjust to many resolutions.

Though depending on varibles. I got unexpected results. It's pretty good as is right now.

Though I removed the black border around things. I'm unsure if that's a desired effect. we can change how it scales pretty easily and return the black border. I was thinking it increased the used screen area.