New version of strings.cs working.
also i moved multichar bytes to ff8stringreference and made it automaticly expand the bytes to multiple bytes.

also moved the namedic tag processing to ff8stringreference and it now checks for a flag set in the stringsbase class to see if it check for the namedic tags.

ff8string does the reading of the strings on the fly. instead of everything at startup. it waits for a direct Value reference or a check of Length.