bool for battle so Commands knows the difference. Added width change for when limitbreak is detected.
added x,y,width,height to dataitem so i could change them easier.

limitId was replaced with a link to the battle command

renamed VisibleInMenu to Available

CurrentFrameNum is public so now i can set that value  in the feild varibles.
added some quick bools to check statuses that could prevent us from using a menu and what not.