victorymenu wip
added partypos to igmdata that tracks where the character is located in the party.

disambigous'ed the + operator so i can add a ff8stringreference to ff8string with out an error. it casts the ff8stringreference to a ff8string. and at that point string will be read from the file so it won't need to keep the reference info anymore.