trimming icons seems to work. WIP
WIP on trimming transparent pixels to fill the entire line of text with an icon. It's working with one texture and one entry.

multi part textures and entrygroups won't work yet.

I added saving to texture handler to confirm i was getting correct results. is easy to disable

I added another scale varible for the icons though unsure if that is correct yet. it seems to work.

Rectangle in entry group is kinda silly as it only tracks the max height and width. Alot of things break if you change that.  I might need to rework it a bit to get entrygroups of count>1 to work with trim.