Update Entry.cs
Two passes one pads the texture by what you are losing with the new x and y. The other removes the padding. I might wanna make it so the second pass doesn't run with 1 entry as some entries are the same just moving something. But if you want to know exactly where something will draw it will be easier if the offset would tend to be 0,0