Multiple changes
I made Input_OKAY return a bool. So the idea is if True it will be ignored by any following checks for input. if false something else can check for that input. I use this to detect when I've gone through the queue of things in victory menu. Switching modes in victory menu and eventually falling back to previous prebattle state.

Tweaked the code for the crisis level. I fixed the non crisis width.

Made music playing a public bool so i can tell if the music is playing before the battle menu starts.

Before crisis level generator was always returning -1 so we fixed that.

Added sound for limit break turn start. and fixed it so sound from ap victory menu isn't playing on the exp screen.

https://youtu.be/2Z-QAErVpjM