added ability to disable loopping of music.
added some code to allow not looping the music. Because sometimes overture would goto main menu a few seconds early and the music would cut off. So just disabling the music lets the song end when it hits the main menu on it's own. I still need to do fulid midi.