Update module_overture_debug.cs
I trying to debug some issues in the overture. so i set a speed value to speed up the entire thing.

Also changed to totalmilliseonds since it's more accurate when you unlocked the framerate the millisecond was returning too small or useless results.

I had wrapped the splash texture into a class. It might not be necessary but should make it easy to add the squaresoft logo later for the 2000 cd release.