Enemy list in target now shows the spawned enemies.
Damageable class tracks things like current hp and status effects.

EnemyInstanceInformation tracks all the stats of the enemies

So thinking of how the kernel and saves are. I have the Enemy class treat the EII as information to calculate stuff with.

I also made some notes of to dos in commands and a comment in mag pool.