added status resistance and devour and card info
since it isn't 1 to 1 for all the different statuses someone can have and the resistances i made a switch case list for everything there is i think.

Elemental still confuses me. but per wiki it seems that the value is 100 less than the one in the system. being 0 is normal chance and -100 is no resistance. 100+ is immune. I think anyway.  might be more to it.