target window to magic pool
I added target window to magic pool.

I had to make the cursor not draw over the top and the arrows in the corners not draw over the top. The arrows are added to all pools automaitcly and by defautl they have the second and third highest zindex. I made it so you could override the draw item method and so now it just skips drawing if the target window is showing.