Update TextureHandler.cs
optimize the loading of pngs' and commented out the old method of loading.

i'd like to skip loading of texturehandler if they are all the same texture. But I don't know exactly how I can know i don't need to till after I scan the path for the filename and that is done in the texturehandler...