fix card face icon
with unlocked framerate milliseconds was returnning values too small to measure so it wouldn't cycle threw the images. switched to totalmilliseconds.

Also changed the debuglobby to set the mode for the them to draw instead of wait. in wait more draw is skipped.

I made the zoom in on icon test varible and +,- keys go from 1f to 100f. though if it gets big enough for part to be off screen it might not draw.