switched some things to slide.
It's a little clearer what we are getting you can say i want this to be at location by x milliseconds. and it happens. previously it was unclear how much time and/or it wasn't based on time.

And now we can see the correct animation regardless of framerate on loading bar, sliding on lgsg and blinking arrows on load screen.