Added color to draw methods
So we can now pass a color to Icons. and entry color. This way if I have a icon and font drawing with a color I can make them use the same colors.

example there is only 1 red real red palette for icons. not really a good faded red. So i have it draw from the white palette and override the color to more match the font which we are doing the same thing.