simplify character/enemy targets to damageables
Made it so instead of two varibles i have 1 array. you can get typeof to check if enemy or character.

Also made it so team laguna characters can be queryed via this[characters]; This will give us the characterdata of the other character.

The name will now be overriden of the character to a team laguna party member.