Strings uses tasks now.
The strings library can now parce the locations of strings all at the same time. instead of waiting for each one before to process.

The actual reading of the string data will be when the string is used.

Also changed the order of the debug message so file is after thread id