Removed Character and VisibleCharacter replaced with Damageable.
Well it's working. Though I'm 90% sure there are bugs introduced because i had to change so many lines of code. But now I can assign an Character, Enemy, or GF to a menu. hoping for only a few bugs :slight_smile:
Work in progress. I mostly wanted to be able to tell commands screen or battle menu here is the enemy. Make it do something!  Like i'm picturing for debugging we can bring up the enemy attacks and choose them and their targets.