Update module_world_debug.cs
forestAdj also effects the inside of the cave... so 0f for now.

i am still unsure why i can walk off the clifft at the mouth of the cave and land on the ground.

Changed invehicle to bdebugcollision because if collision is disabled you can walk under the world with squall super easy.