Extracted ATBTimer and moved it and BattleModeEnum
BattleMode for each damageable is now tracked in damageable instead of in the battlemenu.

Moved ATBTimer to it's own object and put it an instance in the damageable class.

BattleMenus.cs is still responceable from triggering which character's turn it is. but the damageable class changes from charging to charged.

I need to move the update() ops for damageable to a more perminate home. As long as it's only called once per frame should be fine.

I also restructured the battlemenu class a bit so we aren't creating new objects in refresh. should only be updating the objects instead.