Update IGM_LGSG.cs
Choose game screen is nearly done.

I need to make page selection work correctly.
and also animate the blocks flying in from right to left or left to right.
I want to make each block track it's own position and use the page events to know where it needs to be. It will only hide after it's off screen. :D