added init.out support
init out is a default save data file. that contains only a subsection of the data. So anything else needs to be set to some kind a manual default.

This is now set to memory.state after it's loaded so there is atleast something in memory now instead of everything being null.