messing with section5 and refactor a bit.
mostly just testing stuff. I refactored the battledat a bit so i can return null on failure instead of having a broken class with no valid info in it. This showed me my weapon selection code was broken.

I still allow the default costume on characters as theres only 2 versions of that.