Damageable in menus can be set null now.
previously if set null the value wouldn't propogate to the child members. So now i have a bool to force this to happen. You don't always want this because it could cause issues. Because by default damageable will propogate on refresh. if the parent class didn't use damageable it would send null to all the children. So this lets us control it.

I had to do this so i can change menus to null to make them disabled.

renamed init that set damageable to SetDamageable and made it public. Since base.refresh would re popogate the setDamageable to the children and refresh them. It allowed me to set all the damageables then refresh instead of double or tripple refreshes.

Commands should now use base.refresh to update magic and item pools instead of manually setting them.

i setup enemy partypos to be -1 to -9 so you i use partypos to adjust the location of the name hp atb bar.

I have the commands showing for enemy's but I think i should upgrade this to use item pool and magic pool and also enemy command pool.