WIP enemy pools, Magic seems done.
made Player protected as it isn't as useful outside of battlemenus.

instead added GetDamageable to get the value from the current menu.

fixed a couple spots where Player was pointing to an invalid menu. So had to do null checks and range checks to make sure we are okay.

fromc is now damageable so it can be from enemy or from character

moved random init to the static defination so it would get initiated before anything that uses it.

I changed Pages to protected. I wanted to force pages to 1 for when enemy spells count is < rows. though maybe I had another way to do it and forgot. lol.

Magic pool will add enemy's spells that it can cast as an ability first. Then it'll add the spells from Draw list. Though in real game enemy's can't use all the spells in the draw list. I thought it'd be fun for testing.

Enemy's have unlimited cats so i hid the number of magic.