wip on angelo pool
I had to alter the data structure in saves to make the angelo varibles easier to queue. added 2 enums and a dictionary<angelo,byte>. Dictionary will list the points earned for the angelo enum. the angelo enum is all 8 abilities angelo can learn. The other enum is related to several variables one of which is angel wings unlocked.

changed depth of enemy attacks to 1 might need to be 2.

I copied the code from enemy attacks because it's the most basic pool I could think of.

for debug reasons I'll probably provide a list of all of angelo's abilties to the pool so we can test them in the future.
but by default the options should be angelo or angel wing if angel wing is unlocked else skip display.