wip on Combine/Angelo pool
Added Angelo enum tags to abilities related to them. So I can query for them.
Also added ID's the the those kernel classes. I tend to need to reference those sometimes.
Started to add AngeloPool / Combine to Commands. need to finish that.
Moved Angelo enum into it's own file.
Added Angel_Wing to enum but it doesn't serve a purpose other than to tab abilities related to it. There are 2 in the kernel.
Renamed Angelo pool to Combine.