Combine pool should be working pretty good now.
I think i can move on to another part. when we flush things out we can work our way back.

renamed Angelo stuff to Combine.

pool now populates first 2 are what would be in the actual menu.  rest are just the abilitiys related to angelo and angel wing.

added target / name  / id to the struct to make it easy to pull the data out. Also I am using this to replace the Angelo special character. I can't do that in kernel before hand because the name can change.

though there is no target for njGFs maybe it's part of a battle script since it happens on it's own durring the turns in game.