selphie slots class wip
adding support for random from IEnumerable.

Made so public Casts replaced private count in slot. It is a random value.
Also I get the Magic_Data object instead of getting the magicID. So that we don't need to derive that later on.

Added random to Target Party/Enemies. using new extension.

Randomly select target and cast spell with Target.Random class. I used a class instead of a specific bool for each type to allow for expansion.