smoothstep to lerp
I think it was lerp and i changed to smoothstep. I found that smoothstep is nice in some cases. but not all.  Like looping an animation  going from pos 100 to pos 1 smoothstep will not just go from 100 to 1 it'll slowly over a few frames go back to 1. because of the distance. I think