tuple to struct and method replace 3 spots.
Pi is 180 degrees.
3 is 171.887

I was trying to fallow the tuples are confusing being named ITEM1 Item2 etc. There is a new tuple with names in like C# 8 or something. But we aren't using it. So the closest thing I think is a struct. Or maybe a class because tuples can be null.

Then I consolidated 3 chunks of code into 1 method with 4 arguments.