testing crosshair drawing. wip
This code does draw the crosshair. it is pointing at a location in 3d space. I used the indicator's location for testing.  This works but has no z cord. So it's overlaied on top of 3d objects even if i'm pointing at something behind the object. Which would be useful for some things but not all.