Update module_battle_Debug.cs
moved centerofscreen to it's own property

added constained billboard but might use it.

moved the crosshair down. might be an issue with the cross hair. might have to use the other method to draw it.  where I'm drawing it with sprite batch. to do this I'll need to make battle draw it after drawing the monster or character.