Moved DrawCrossHair into module_battle_debug.
Because Spritebatch will draw over the monster always i figured this would be better than the quad method. because it's hard to make the quad not overlap with the targets.

This basically draws the monster than checks for a crosshair and draws it if needed. Though the code is a bit messy. :P