added check for 8bpp and force number of colors to 16 if above that number.
I saw timviewer was forcing 4bpp tims to 16 colors per pixel. Where my code was assuming the width was the number I needed. And so I now force it to 16 if it is above 16 . I think it's possible that someone could want less color's per pixel but it cannot be greater than 16.