Applied changes to archiveworker to fields. And tweaked color code for bg texture.
So I made some changes to where it reads in the data from the archive worker. I moved the byte[]'s into archiveworker and gave us an easy way to request them.
I'm not sure if it's all read correctly. It bothers me I see black spots on backgrounds.

I kinda wonder if we should be drawing this on quads instead of sprite batch. we could put the over lapping tiles into different layers with slightly different z coords. something like that.

I would like to compare how we render to how other software does it.