Rename overlap to Intersect and added a rectangle to store where the two intersect
This is leading to a pixel perfect overlap detection which requires i can test the color when i'm incrementing the overlapID.

So basically I want to check the colors for transparent. so not both can have color at the same time. if they are a perfect cutout then we merge together.