Fixed bad merge and fixed errors caused by my changes.
ArchiveWorker can now read an internal archive and return it as another archiveworker/archivebase. I setup a base class to support reading from other filetypes with the same interface. I might go back to an interface. But I thought an abstract class could be useful if I am going to use shared code. Though right now i'm not.

So I changed the code in field_object_test to GetArchive() instead of requesting and keeping track of each of the 3 files.