Update module_field_object_test.cs
ViewPort is correct aspect ratio to use.  DisplayMode is based on your monitor. so if you change monitors your pespective could change.

I was testing alphablend. I was thinking if her glasses are transparent black it should of made it so we could see through em. But it didn't work.