Added some code to reduce the amount of palette dumps.
though this code can slow some things down so only used in reswizzle.

the way it works is it checks for overlap when it detects you are trying to overwrite a pixel. Flags all the palette ids. and only processes the palette ids that are flagged when redoing the work.