camera and stage seem to be a a better place.
goal was to group up the stage values and camera values into their own classes. Debug module was very cluttered and it was hard to find anything. And everything seems to be working.
I also wanted camera to not be stuck in the stage so module had control over it. So I just pass the endoffset from camera into stage. Though one issue with this approch is the archive file is opened multiple times. I'll probably move the open method to module. Though when camera changes animations it re opens to read the next one instead of storing all of them in an array.