Changed MSD Reader to use FF8String and BinaryReader.
FF8String keeps the strings encoded in byte[] till we need a unicode version. Or font renderer is expencting bytes. when it tries to draw. And his conversion to string includes alot of bracketed strings to signify what 1 or 2 bytes will. I don't think there is really a reason to use unsafe code when we aren't using third party libraries that force us to.