tweaked the code for encounter
changed elements to a bitarray.
added location to EII and CII to so other objects can reference the lcoation.
I was thinking that we might not of been pulling the correct location for enemies. so wanted to assign it when the enemy was loaded. And not to the dat class as it is shared between enemies of the same type.