Update ArchiveWorker.cs
Tried to make ArchiveWorker work like a Dictionary for foreach. For those times you will be loading all the files.

corrected a change i made earlier. Where i assumed the FI was in the same order as the FS file. So i used the following location to calculate the length. of the file. And it seems to not be. Might be worth putting the FI and FL information together. depends if Lz4 cares about haveing too much data.