moved buffer_data to av folder as I plan on using it to read from other.zzz after tweaks
I kinda have a over arching plan to read from zzz files with out just dumping the entire contents into ram. using the other.zzz to know how to open the stream and be able to dump the offset and size if need be.

am dumping the ogg file names from from other.zzz so i know where they are and they show in the menus.

this hurts my brain. heh.