i ment to be 5 or greater.
I kinda wish i could detect a valid compressed file size. but as far as I know there isn't a way. But fzss seems to work with any number of bytes. but if you give it invalid data you get invalid data back out. And the odds of getting the exact filesize from it are low i think if the file is invalid.