fix back button on field debug menu
I was relying on a  global back button = go to main menu. so this cause issues when running in release mod that didn't have this global. I changed the global to only trigger on Reset combo. like on psx and pc theres a button combo triggers reset.

made a reload texture method to make it clear what was happening in code.

I added the code to escape to menu a few other places that were missing it. control R is a shortcut if you get stuck somewhere.