Fix reswizzle (#186)
* Create Save_As_PNG function

This is to wrap the FileStream Create in a try catch block.  It logs exceptions on error. Before if it error'ed it would crash.

* Use Save As PNG function

Change code that save's pngs to use the new function

* Expanded Source was wrong.

Expanded Source would double X if it was 4bit. And it didn't work. I just didn't notice till now. So I commented it out and Corrected code to not use it.