Dialog box works now to replace exception.
This is what the replacement looks like. If you press ignore or retry it'll just keep replacing any pixel that conflicts. I might need to check code. But this might be good enough. I think it's tile(s) that are used in multiple places so you guys don't have consistent colors

https://twitter.com/sebanisu/status/1397269933680115712